Feature/visualitzar detalls ruta

See merge request ymelero/entrebicis!20

+++ FrontEnd/app/src/main/java/cat/copernic/ymelero/entrebicis/core/model/Recompensa.kt
@@ -4,7 +4,7 @@ data class Recompensa (
     val id: Long,
     val descripcio: String,
     val observacions: String?,
-    val punts: Int,
+    val punts: Double,
     val estat: EstatRecompensa,
     val nomComerc: String,
     val adrecaComerc: String,
